bugfix cyclic mapper and added precondition in changeset

+++ src/main/java/ru/vadim/hostel/repository/GuestRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface GuestRepository extends JpaRepository<Guest, Long> {
 
-    Optional<Guest> findGuestByPassport(Integer passportNumber);
+    Optional<Guest> findGuestByPassport(Long passportNumber);
 }
